+++ framework/server/src/test/java/is/codion/framework/server/DefaultEntityServerAdminTest.java
@@ -25,7 +25,6 @@ public final class DefaultEntityServerAdminTest {
   void test() throws Exception {
     Clients.SERVER_HOSTNAME.set("localhost");
     Clients.TRUSTSTORE.set("src/main/config/truststore.jks");
-    Clients.TRUSTSTORE_PASSWORD.set("crappypass");
     Clients.resolveTrustStore();
     ServerConfiguration.RMI_SERVER_HOSTNAME.set("localhost");
     ServerConfiguration.KEYSTORE.set("src/main/config/keystore.jks");